+++ src/App.js
@@ -1,20 +1,18 @@
 import './App.css';
 import { ChakraProvider } from '@chakra-ui/react'
-import LeftLogin from './LoginComponents/LeftLogin';
-import RightLogin from './LoginComponents/RightLogin';
+import LeftLogin from './Login/Component/LeftLogin';
+import RightLogin from './Login/Component/RightLogin';
 import { HStack } from '@chakra-ui/react'
 
 function App() {
 
   return (
-    <div className="App">
-      <ChakraProvider >
-        <HStack>
-          <LeftLogin />
-          <RightLogin />
-        </HStack>
-      </ChakraProvider>
-    </div>
+    <ChakraProvider >
+      <HStack>
+        <LeftLogin />
+        <RightLogin />
+      </HStack>
+    </ChakraProvider>
   );
 }
